requirements.txt
acvl_utils==0.2
alembic==1.13.1
asgiref==3.8.1
batchgenerators==0.25
certifi==2024.2.2
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
connected-components-3d==3.12.5
contextlib2==21.6.0
contourpy==1.2.0
crashtest==0.4.1
cycler==0.12.1
dataclasses==0.6
dcmrtstruct2nii==5
decorator==5.1.1
dicom2nifti==2.4.10
Django==5.0.3
dynamic_network_architectures==0.3.1
filelock==3.13.3
fonttools==4.50.0
fsspec==2024.3.1
future==1.0.0
graphviz==0.20.3
greenlet==3.0.3
idna==3.6
imagecodecs==2024.1.1
imageio==2.34.0
Jinja2==3.1.3
joblib==1.3.2
kiwisolver==1.4.5
lazy_loader==0.3
linecache2==1.0.0
Mako==1.3.2
MarkupSafe==2.1.5
matplotlib==3.8.3
MedPy==0.4.0
mpmath==1.3.0
mysqlclient==2.2.4
networkx==3.2.1
nibabel==5.2.1
nnunet==1.7.1
nnunetv2==2.3.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
opencv-python==4.9.0.80
packaging==24.0
pandas==2.2.1
pillow==10.2.0
py==1.11.0
pydicom==2.4.4
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-gdcm==3.0.23
pytz==2024.1
PyYAML==6.0
rapidfuzz==3.7.0
requests==2.31.0
retry==0.9.2
rt-utils==1.2.7
schema==0.7.5
scikit-image==0.22.0
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
SimpleITK==2.3.1
six==1.16.0
SQLAlchemy==2.0.29
sqlparse==0.4.4
sympy==1.12
termcolor==2.4.0
threadpoolctl==3.4.0
tifffile==2024.2.12
tk==0.1.0
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tqdm==4.66.2
traceback2==1.4.0
triton==2.2.0
typing_extensions==4.10.0
tzdata==2024.1
unittest2==1.1.0
urllib3==2.2.1
watchdog==3.0.0
yacs==0.1.8
